refactor: move sortByStartsAt
