+++ src/hooks/useAllEventsData.ts
@@ -2,26 +2,10 @@ import { CREATE_EVENT_SCHEMA_UID } from "@/hooks/useCreateEventAttestation";
 import { useSigner } from "@/hooks/useSigner";
 import { API_URL_MAPPING } from "@/lib/gqlEasAttestation";
 import { EVENTS_ATTESTATIONS_QUERY } from "@/lib/gqlEasAttestation/query";
-import type { DataEntry } from "@/types";
+import { sortByStartsAt } from "@/lib/utils";
 import { useMemo } from "react";
 import { useQuery } from "urql";
 import { useChainId } from "wagmi";
-
-function sortByStartsAt(data: DataEntry[]): DataEntry[] {
-  return data.sort((a, b) => {
-    const aDecoded = JSON.parse(a.decodedDataJson);
-    const bDecoded = JSON.parse(b.decodedDataJson);
-
-    const aStartsAt = aDecoded.find((item: any) => item.name === "startsAt")
-      ?.value.value.hex;
-    const bStartsAt = bDecoded.find((item: any) => item.name === "startsAt")
-      ?.value.value.hex;
-
-    if (!aStartsAt || !bStartsAt) return 0;
-
-    return Number.parseInt(aStartsAt, 16) - Number.parseInt(bStartsAt, 16);
-  });
-}
 
 export const useAllEventsData = () => {
   const signer = useSigner();
